requirements.txt
asgiref==3.8.1
attrs==24.2.0
dj-database-url==2.2.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
